Simplify data source creation in database migration
