+++ multiapps-controller-database-migration/src/main/java/org/cloudfoundry/multiapps/controller/database/migration/extractor/DataSourceEnvironmentExtractor.java
@@ -2,43 +2,29 @@ package org.cloudfoundry.multiapps.controller.database.migration.extractor;
 
 import javax.sql.DataSource;
 
-import org.postgresql.ds.PGSimpleDataSource;
+import org.cloudfoundry.multiapps.controller.persistence.util.DataSourceFactory;
+import org.cloudfoundry.multiapps.controller.persistence.util.EnvironmentServicesFinder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.pivotal.cfenv.core.CfCredentials;
-import io.pivotal.cfenv.core.CfEnv;
-import io.pivotal.cfenv.core.CfService;
+import io.pivotal.cfenv.jdbc.CfJdbcService;
 
 public class DataSourceEnvironmentExtractor {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(DataSourceEnvironmentExtractor.class);
 
     public DataSource extractDataSource(String serviceName) {
-        LOGGER.info("Extracting datasource for service {}...", serviceName);
-        CfCredentials databaseServiceCredentials = extractDatabaseServiceCredentials(serviceName);
-        return extractDataSource(databaseServiceCredentials);
+        LOGGER.info("Creating a data source for service {}...", serviceName);
+        CfJdbcService service = findService(serviceName);
+        return createDataSource(service);
     }
 
-    private CfCredentials extractDatabaseServiceCredentials(String serviceName) {
-        CfEnv cfEnv = new CfEnv();
-        CfService sourceService = cfEnv.findServiceByName(serviceName);
-        return sourceService.getCredentials();
+    private CfJdbcService findService(String serviceName) {
+        return new EnvironmentServicesFinder().findJdbcService(serviceName);
     }
 
-    private PGSimpleDataSource extractDataSource(CfCredentials databaseServiceCredentials) {
-        PGSimpleDataSource dataSource = new PGSimpleDataSource();
-        dataSource.setServerNames(new String[] { databaseServiceCredentials.getHost() });
-        dataSource.setUser(databaseServiceCredentials.getUsername());
-        dataSource.setPassword(databaseServiceCredentials.getPassword());
-        dataSource.setDatabaseName(databaseServiceCredentials.getString("dbname"));
-        dataSource.setPortNumbers(new int[] { getPort(databaseServiceCredentials) });
-        dataSource.setSsl(false);
-        return dataSource;
-    }
-
-    private int getPort(CfCredentials databaseServiceCredentials) {
-        return Integer.valueOf(databaseServiceCredentials.getPort());
+    private DataSource createDataSource(CfJdbcService service) {
+        return new DataSourceFactory().createDataSource(service);
     }
 
 }
