+++ multiapps-controller-persistence/src/main/java/org/cloudfoundry/multiapps/controller/persistence/util/EnvironmentServicesFinder.java
@@ -1,4 +1,4 @@
-package org.cloudfoundry.multiapps.controller.web.util;
+package org.cloudfoundry.multiapps.controller.persistence.util;
 
 import javax.inject.Named;
