+++ multiapps-controller-web/src/main/java/org/cloudfoundry/multiapps/controller/web/configuration/DatabaseConfiguration.java
@@ -7,9 +7,9 @@ import org.cloudfoundry.multiapps.controller.core.util.ApplicationConfiguration;
 import org.cloudfoundry.multiapps.controller.persistence.DataSourceWithDialect;
 import org.cloudfoundry.multiapps.controller.persistence.dialects.DataSourceDialect;
 import org.cloudfoundry.multiapps.controller.persistence.dialects.DefaultDataSourceDialect;
+import org.cloudfoundry.multiapps.controller.persistence.util.DataSourceFactory;
+import org.cloudfoundry.multiapps.controller.persistence.util.EnvironmentServicesFinder;
 import org.cloudfoundry.multiapps.controller.web.configuration.bean.factory.CloudDataSourceFactoryBean;
-import org.cloudfoundry.multiapps.controller.web.configuration.bean.factory.DataSourceFactory;
-import org.cloudfoundry.multiapps.controller.web.util.EnvironmentServicesFinder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.DependsOn;
