+++ multiapps-controller-web/src/test/java/org/cloudfoundry/multiapps/controller/web/configuration/bean/factory/CloudDataSourceFactoryBeanTest.java
@@ -2,14 +2,15 @@ package org.cloudfoundry.multiapps.controller.web.configuration.bean.factory;
 
 import static org.junit.Assert.assertEquals;
 
-import org.cloudfoundry.multiapps.controller.web.util.EnvironmentServicesFinder;
+import javax.sql.DataSource;
+
+import org.cloudfoundry.multiapps.controller.persistence.util.DataSourceFactory;
+import org.cloudfoundry.multiapps.controller.persistence.util.EnvironmentServicesFinder;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
-
-import com.zaxxer.hikari.HikariDataSource;
 
 import io.pivotal.cfenv.jdbc.CfJdbcService;
 
@@ -18,7 +19,7 @@ public class CloudDataSourceFactoryBeanTest {
     private static final String SERVICE_NAME = "my-service";
 
     @Mock
-    private HikariDataSource dataSource;
+    private DataSource dataSource;
     @Mock
     private DataSourceFactory dataSourceFactory;
     @Mock
@@ -37,7 +38,7 @@ public class CloudDataSourceFactoryBeanTest {
         CfJdbcService service = Mockito.mock(CfJdbcService.class);
         Mockito.when(vcapServiceFinder.findJdbcService(SERVICE_NAME))
                .thenReturn(service);
-        Mockito.when(dataSourceFactory.createHikariDataSource(service))
+        Mockito.when(dataSourceFactory.createDataSource(service))
                .thenReturn(dataSource);
 
         testedFactory.afterPropertiesSet();
