+++ pom.xml
@@ -49,7 +49,6 @@
         <resilience4j.version>1.5.0</resilience4j.version>
         <immutables.version>2.8.8</immutables.version>
         <micrometer.version>1.5.4</micrometer.version>
-        <postgresql.version>42.2.9</postgresql.version>
         <aliyun-sdk-oss.version>3.11.0</aliyun-sdk-oss.version>
         <auto-service.version>1.0-rc7</auto-service.version>
         <java-cfenv.version>2.2.2.RELEASE</java-cfenv.version>
@@ -778,13 +777,6 @@
                 <groupId>io.github.resilience4j</groupId>
                 <artifactId>resilience4j-ratelimiter</artifactId>
                 <version>${resilience4j.version}</version>
-            </dependency>
-            <!-- https://mvnrepository.com/artifact/org.postgresql/postgresql -->
-            <dependency>
-                <groupId>org.postgresql</groupId>
-                <artifactId>postgresql</artifactId>
-                <scope>provided</scope>
-                <version>${postgresql.version}</version>
             </dependency>
             <!-- https://mvnrepository.com/artifact/io.pivotal.cfenv/java-cfenv -->
             <dependency>
